Separating update command builder functionality into separate modules by entity type
